crate skeleton, move to validation
