+++ src/rust/cryptography-x509-validation/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cryptography-x509-validation"
+version = "0.1.0"
+authors = ["The cryptography developers <cryptography-dev@python.org>"]
+edition = "2021"
+publish = false
+# This specifies the MSRV
+rust-version = "1.56.0"
+
+[dependencies]
+asn1 = { version = "0.15.0", default-features = false }
+cryptography-x509 = { path = "../cryptography-x509" }
+
+[dev-dependencies]
+pem = "1.1"
