+++ src/rust/cryptography-x509-validation/src/lib.rs
@@ -1,0 +1,7 @@
+// This file is dual licensed under the terms of the Apache License, Version
+// 2.0, and the BSD License. See the LICENSE file in the root of this repository
+// for complete details.
+
+#![forbid(unsafe_code)]
+
+pub mod types;
